requirements.txt
openai>=1.0.0
langchain>=0.2.0
langchain-community>=0.0.15
langchain-openai>=0.0.3
faiss-cpu>=1.7.4
pandas>=1.5.0
tiktoken>=0.5.1
python-dotenv>=1.0.0
colorama